+++ src/FileSystem/storage.ts
@@ -2,6 +2,7 @@ import { RepoNode } from "../Tree/nodes";
 import { ExtensionContext } from "vscode";
 import { GLOBAL_STORAGE_KEY } from "../GitHub/constants";
 import { output, repoProvider } from "../extension";
+import { openRepository } from "../GitHub/commands";
 
 export const store = {
     repos: [] as (RepoNode | undefined)[],
@@ -14,15 +15,15 @@ export const store = {
  * @param {ExtensionContext} context Extension context
  * @param {string} value Repository to add
  */
-export function addToGlobalStorage(context: ExtensionContext, value: string): void {
-    let globalStorage: string[] = getReposFromGlobalStorage(context);
+export async function addToGlobalStorage(context: ExtensionContext, value: string): Promise<void> {
+    let globalStorage = await getReposFromGlobalStorage(context);
     globalStorage.push(value);
     context.globalState.update(GLOBAL_STORAGE_KEY, globalStorage);
 
     repoProvider.refresh();
 
-    output.appendLine(`Added ${value} to global storage`, output.messageType.info);
-    output.appendLine(`Global storage: ${globalStorage}`, output.messageType.info);
+    output?.appendLine(`Added ${value} to global storage`, output.messageType.info);
+    output?.appendLine(`Global storage: ${globalStorage}`, output.messageType.info);
 }
 
 /**
@@ -40,8 +41,8 @@ export function removeFromGlobalStorage(context: ExtensionContext, value: string
 
         repoProvider.refresh();
 
-        output.appendLine(`Removed ${value} from global storage`, output.messageType.info);
-        output.appendLine(`Global storage: ${globalStorage}`, output.messageType.info);
+        output?.appendLine(`Removed ${value} from global storage`, output.messageType.info);
+        output?.appendLine(`Global storage: ${globalStorage}`, output.messageType.info);
     }
 }
 
@@ -52,12 +53,37 @@ export function removeFromGlobalStorage(context: ExtensionContext, value: string
  * @param {ExtensionContext} context Extension context
  * @returns {string[]}
  */
-export function getReposFromGlobalStorage(context: ExtensionContext): string[] {
+export async function getReposFromGlobalStorage(context: ExtensionContext): Promise<string[]> {
     return context.globalState.get(GLOBAL_STORAGE_KEY, []);
+    // return await purgeGlobalStorage(context);
 }
 
-export function clearGlobalStorage(context: ExtensionContext) { 
+export function clearGlobalStorage(context: ExtensionContext) {
     context.globalState.update(GLOBAL_STORAGE_KEY, []);
-    output.appendLine(`Cleared global storage`, output.messageType.info);
+    output?.appendLine(`Cleared global storage`, output.messageType.info);
     repoProvider.refresh();
+}
+
+export async function purgeGlobalStorage(context: ExtensionContext, repos?: string[]): Promise<string[]> {
+    let cleanedGlobalStorage: string[] = [];
+    if (repos) {
+        cleanedGlobalStorage = repos.filter((item) => item !== undefined);
+        context.globalState.update(GLOBAL_STORAGE_KEY, cleanedGlobalStorage);
+    } else {
+        const globalStorage = context.globalState.get(GLOBAL_STORAGE_KEY, []) as string[];
+        cleanedGlobalStorage = await Promise.all(
+            globalStorage.map(async (repo) => {
+                let repoOwner = repo.split("/")[0];
+                let repoName = repo.split("/")[1];
+                let validRepo = await openRepository(repoOwner, repoName);
+                if (validRepo) {
+                    return Promise.resolve(`${validRepo.owner.login}/${validRepo.name}`);
+                } else {
+                    return Promise.reject();
+                }
+            })
+        );
+    }
+
+    return cleanedGlobalStorage;
 }